Here are my attempts at the Labs for Week 1
One screenshot has errors in the console. My code was similar to the videos, yet when I try to run with the variable names for Part 2, it gave me those errors. 
So in the end, I re-wrote the names of the variables for Part 2 of the Lab. 
Anyone care to elaborate? 
Cheers